pyproject.toml
[tool.poetry]
name = "flask-web-socket"
version = "0.1.0"
description = ""
authors = ["leoneville <leoneville_@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "2.3.0"
qrcode = "7.4.2"
pillow = "10.2.0"
flask-socketio = "5.3.6"
flask-sqlalchemy = "3.1.1"
flask-cors = "^4.0.1"
pytest = "^8.2.1"
mercadopago = "^2.2.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
